+++ packages/editor-core/src/editor/create-editor/get-props-preset.ts
@@ -1,0 +1,23 @@
+import { EditorProps, EditorAppearance } from '../types';
+
+const messageEditorPropsPreset: EditorProps = {
+  appearance: 'message',
+
+  saveOnEnter: true,
+
+  allowTextFormatting: true,
+  allowLists: true,
+  allowTextColor: true,
+  allowHyperlinks: true,
+  allowCodeBlocks: true,
+  allowTasksAndDecisions: true
+};
+
+export default function getPropsPreset(appearance: EditorAppearance): EditorProps {
+  switch (appearance) {
+    case 'message':
+      return { ...messageEditorPropsPreset };
+    default:
+      return {};
+  }
+}