+++ packages/editor-core/src/editor/index.tsx
@@ -5,6 +5,7 @@ import { createPluginsList } from './create-editor';
 import EditorActions from './actions';
 import ProviderFactory from '../providerFactory';
 import { EditorProps, EditorInstance, EditorAppearanceComponentProps } from './types';
+import { moveCursorToTheEnd } from '../utils';
 export * from './types';
 
 export interface State {
@@ -82,17 +83,30 @@ export default class Editor extends React.Component<EditorProps, State> {
     this.setState({ editor });
 
     // Focus editor first time we create it if shouldFocus prop is set to true.
-    if (this.props.shouldFocus && !editor.editorView.hasFocus()) {
-      editor.editorView.focus();
+    if (this.props.shouldFocus) {
+      if (!editor.editorView.hasFocus()) {
+        editor.editorView.focus();
+      }
+
+      moveCursorToTheEnd(editor.editorView);
     }
   }
 
   private handleProviders(props: EditorProps) {
-    const { emojiProvider, mentionProvider, mediaProvider, collabEditProvider } = props;
+    const {
+      emojiProvider,
+      mentionProvider,
+      mediaProvider,
+      collabEditProvider,
+      activityProvider,
+      presenceProvider
+    } = props;
     this.providerFactory.setProvider('emojiProvider', emojiProvider);
     this.providerFactory.setProvider('mentionProvider', mentionProvider);
     this.providerFactory.setProvider('mediaProvider', mediaProvider);
     this.providerFactory.setProvider('collabEditProvider', collabEditProvider);
+    this.providerFactory.setProvider('activityProvider', activityProvider);
+    this.providerFactory.setProvider('presenceProvider', presenceProvider);
   }
 
   render() {
@@ -120,6 +134,8 @@ export default class Editor extends React.Component<EditorProps, State> {
 
         eventDispatcher={eventDispatcher}
 
+        maxHeight={this.props.maxHeight}
+
         contentComponents={contentComponents}
         primaryToolbarComponents={primaryToolbarComponents}
         secondaryToolbarComponents={secondaryToolbarComponents}
@@ -127,6 +143,7 @@ export default class Editor extends React.Component<EditorProps, State> {
         customContentComponents={this.props.contentComponents}
         customPrimaryToolbarComponents={this.props.primaryToolbarComponents}
         customSecondaryToolbarComponents={this.props.secondaryToolbarComponents}
+
         addonToolbarComponents={this.props.addonToolbarComponents}
       />
     );