+++ packages/editor-core/src/test-helper/index.ts
@@ -23,3 +23,4 @@ export { default as storyDecorator } from './story-decorator';
 export { storyMediaProviderFactory, getLinkCreateContextMock } from './media-provider';
 export { default as randomId } from './random-id';
 export { default as sleep } from './sleep';
+export { isMobileBrowser } from './device';