+++ packages/editor-core/test/browser/renderer/react/nodes/taskItem.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { mount } from 'enzyme';
+import { mount, shallow } from 'enzyme';
 import { expect } from 'chai';
 import { ResourcedTaskItem as AkTaskItem } from '@atlaskit/task-decision';
 import TaskItem from '../../../../../src/renderer/react/nodes/taskItem';
@@ -9,5 +9,10 @@ describe('Renderer - React/Nodes/TaskItem', () => {
 
   it('should wrap content with <AkTaskItem>-tag', () => {
     expect(taskItem.find(AkTaskItem).length).to.equal(1);
+  });
+
+  it('should not render if no children', () => {
+    const taskItem = shallow(<TaskItem localId="task-2"/>);
+    expect(taskItem.isEmptyRender()).to.equal(true);
   });
 });