+++ packages/emoji/src/components/common/LoadingEmojiComponent.tsx
@@ -13,29 +13,43 @@ export interface State {
 /**
  * A base class for components that don't want to start rendering
  * until the EmojiProvider is resolved.
+ * Notes: super.componentDidMount and super.componentWillUnmount will need to be
+ * called explicitly if they are overridden on the child class.
  */
 export default abstract class LoadingEmojiComponent<P extends Props,S extends State> extends PureComponent<P,S> {
+  private isUnmounted: boolean;
 
   constructor(props: P, state: S) {
     super(props);
     this.state = state;
+  }
+
+  componentDidMount() {
     this.loadEmojiProvider(this.props.emojiProvider);
   }
 
-  componentWillReceiveProps(nextProps: Props) {
+  componentWillReceiveProps(nextProps: Readonly<P>) {
     this.loadEmojiProvider(nextProps.emojiProvider);
   }
 
   private loadEmojiProvider(futureEmojiProvider: Promise<EmojiProvider>) {
     futureEmojiProvider.then(loadedEmojiProvider => {
-      this.setState({
-        loadedEmojiProvider
-      });
+      if (!this.isUnmounted) {
+        this.setState({
+          loadedEmojiProvider
+        });
+      }
     }).catch(err => {
-      this.setState({
-        loadedEmojiProvider: undefined
-      });
+      if (!this.isUnmounted) {
+        this.setState({
+          loadedEmojiProvider: undefined
+        });
+      }
     });
+  }
+
+  componentWillUnmount() {
+    this.isUnmounted = true;
   }
 
   renderLoading(): JSX.Element | null {