+++ packages/layer-manager/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './components/LayerManager';
+export { default as withRenderTarget } from './components/withRenderTarget';
+export { Gateway, GatewayDest, GatewayProvider, GatewayRegistry } from './components/gateway';