+++ packages/size-detector/stories/components/StoryPadding.jsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export default styled.div`
+  padding: 60px;
+  height: 100%;
+  box-sizing: border-box;
+`;