+++ packages/task-decision/src/styled/Placeholder.tsx
@@ -6,7 +6,7 @@ import {
 
 // tslint:disable-next-line:variable-name
 export const Placeholder = styled.span`
-  margin: 1px 0 1px ${akGridSizeUnitless * 4}px;
+  margin: 0 0 0 ${akGridSizeUnitless * 3}px;
   position: absolute;
   color: ${akColorN80};
   pointer-events: none;