+++ packages/task-decision/stories/TaskList-story.tsx
@@ -1,5 +1,6 @@
 import { action, storiesOf } from '@kadira/storybook';
 import * as React from 'react';
+import { PureComponent } from 'react';
 import { ReactRenderer as Renderer } from '@atlaskit/editor-core/dist/es5/renderer';
 
 import TaskList from '../src/components/TaskList';
@@ -11,22 +12,55 @@ const dumpRef = (ref: HTMLElement) => {
   console.log('Content HTML', ref && ref.outerHTML);
 };
 
+interface Props {
+  render: (taskStates: Map<string, boolean>, onChangeListener: (taskId: string, done: boolean) => void) => JSX.Element;
+}
+
+interface State {
+  tick: number;
+}
+
+class TaskStateManager extends PureComponent<Props,State> {
+  private taskStates = new Map<string, boolean>();
+
+  constructor(props) {
+    super(props);
+    this.state = {
+      tick: 0
+    };
+  }
+
+  private onChangeListener = (taskId, done) => {
+    action('onChange');
+    this.taskStates.set(taskId, done);
+    this.setState({ tick: this.state.tick + 1 });
+  }
+
+  render() {
+    return this.props.render(this.taskStates, this.onChangeListener);
+  }
+}
+
 storiesOf('<TaskList/>', module)
   .add('Simple TaskList', () => (
-    <TaskList>
-      <TaskItem contentRef={dumpRef} taskId="task-1" onChange={action('onChange')}>
-        Hello <b>world</b>.
-      </TaskItem>
-      <TaskItem contentRef={dumpRef} taskId="task-2" onChange={action('onChange')}>
-        <Renderer document={document} />
-      </TaskItem>
-      <TaskItem contentRef={dumpRef} taskId="task-3" onChange={action('onChange')}>
-        Hello <b>world</b>.
-      </TaskItem>
-      <TaskItem contentRef={dumpRef} taskId="task-4" onChange={action('onChange')}>
-        <Renderer document={document} />
-      </TaskItem>
-    </TaskList>
+    <TaskStateManager
+      render={(taskStates, onChangeListener) => // tslint:disable-line:jsx-no-lambda
+        <TaskList>
+          <TaskItem contentRef={dumpRef} taskId="task-1" onChange={onChangeListener} isDone={taskStates.get('task-1')} >
+            Hello <b>world</b>.
+          </TaskItem>
+          <TaskItem contentRef={dumpRef} taskId="task-2" onChange={onChangeListener} isDone={taskStates.get('task-2')}>
+            <Renderer document={document} />
+          </TaskItem>
+          <TaskItem contentRef={dumpRef} taskId="task-3" onChange={onChangeListener} isDone={taskStates.get('task-3')}>
+            Hello <b>world</b>.
+          </TaskItem>
+          <TaskItem contentRef={dumpRef} taskId="task-4" onChange={onChangeListener} isDone={taskStates.get('task-4')}>
+            <Renderer document={document} />
+          </TaskItem>
+        </TaskList>
+     }
+    />
   ))
   .add('Single item TaskList', () => (
     <TaskList>