+++ packages/xflow/stories/providers/MockConfluenceXFlowProvider.jsx
@@ -22,7 +22,6 @@ const overrideImplementations = {
   grantAccessToUsers: notImplemented,
   retrieveUsers: notImplemented,
   goToProduct: notImplemented,
-  goToLearnMore: notImplemented,
   closeLoadingDialog: notImplemented,
 };