+++ packages/editor-core/docs/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @atlaskit/editor-core
 
+## 46.14.1 (2017-09-05)
+
+* bug fix; we need to make sure the component is not unmounted before we use this.setState in p (issues closed: ed-2448) ([b3301ea](https://bitbucket.org/atlassian/atlaskit/commits/b3301ea))
+
+* bug fix; tolerate empty actions/decisions, just don't render them. (issues closed: fs-1266) ([29a82db](https://bitbucket.org/atlassian/atlaskit/commits/29a82db))
+* bug fix; eNTER on empty action/decision will not submit message, but remove action/decision (issues closed: ed-2447 / fs-1265 / fs-1329) ([9f77faa](https://bitbucket.org/atlassian/atlaskit/commits/9f77faa))
+* bug fix; disabling hyperlink keymap Cmd-K for message editor. (issues closed: #ed-2567) ([ddf7769](https://bitbucket.org/atlassian/atlaskit/commits/ddf7769))
+## 46.14.0 (2017-09-05)
+
+* feature; add showSpinner prop to Editor ([ac14e79](https://bitbucket.org/atlassian/atlaskit/commits/ac14e79))
+* feature; add defaultValue property for the "new arch" editor (issues closed: ed-2579) ([dc537cc](https://bitbucket.org/atlassian/atlaskit/commits/dc537cc))
+* feature; add appendText command to EditorActions (issues closed: ed-2579) ([417f348](https://bitbucket.org/atlassian/atlaskit/commits/417f348))
+* feature; add waitForMediaUpload flag for products that want to finalize media on their own (issues closed: ed-2579) ([7ee8106](https://bitbucket.org/atlassian/atlaskit/commits/7ee8106))
+* feature; support presence provider ([4926ae0](https://bitbucket.org/atlassian/atlaskit/commits/4926ae0))
+* bug fix; shouldFocus must place cursor to the end of the document (issues closed: ed-2579) ([e3f7187](https://bitbucket.org/atlassian/atlaskit/commits/e3f7187))
+* feature; add maxHeight property to the 'new architecture' editor (issues closed: ed-2579) ([0cad06d](https://bitbucket.org/atlassian/atlaskit/commits/0cad06d))
+* feature; add feature preset for Message editor (issues closed: ed-2579) ([69c5ee1](https://bitbucket.org/atlassian/atlaskit/commits/69c5ee1))
+## 46.13.3 (2017-09-04)
+
+* bug fix; fix filmstrip arrows don't show up in editor or renderer (issues closed: ed-2577) ([43291f6](https://bitbucket.org/atlassian/atlaskit/commits/43291f6))
+## 46.13.2 (2017-09-01)
+
+* bug fix; pass eventDispatcher to PluginSlot (issues closed: ed-2634) ([ee59df4](https://bitbucket.org/atlassian/atlaskit/commits/ee59df4))
+* bug fix; fix editor-core storybook for real (issues closed: ed-2614) ([704a76e](https://bitbucket.org/atlassian/atlaskit/commits/704a76e))
+* bug fix; do not show frequent links in the recently viewed link dialog ([5a175e9](https://bitbucket.org/atlassian/atlaskit/commits/5a175e9))
+* bug fix; fire onChange only when docChanged = true (issues closed: ed-2421) ([5605a74](https://bitbucket.org/atlassian/atlaskit/commits/5605a74))
+## 46.13.1 (2017-09-01)
+
+* bug fix; support activity provider in hyperlinkedit in new editor architecture ([d45ee4e](https://bitbucket.org/atlassian/atlaskit/commits/d45ee4e))
+## 46.13.0 (2017-09-01)
+
+* feature; pasting a single word or single line of code should create an inline code block. ([e64a721](https://bitbucket.org/atlassian/atlaskit/commits/e64a721))
+* bug fix; adjust behaviour of backspace into a decision/item list (issues closed: fs-1264) ([5db2517](https://bitbucket.org/atlassian/atlaskit/commits/5db2517))
+## 46.12.0 (2017-08-31)
+
+* bug fix; update min-height of single link card. ([bb7e71a](https://bitbucket.org/atlassian/atlaskit/commits/bb7e71a))
+* bug fix; fix storybook for editor-core (issues closed: ed-2614) ([4bae650](https://bitbucket.org/atlassian/atlaskit/commits/4bae650))
+
+* bug fix; making sure adding links with invalid href doesn't throw an error (issues closed: ed-2568) ([d44226f](https://bitbucket.org/atlassian/atlaskit/commits/d44226f))
+* bug fix; fixed the bug that no text format after space. ([8aa1bfe](https://bitbucket.org/atlassian/atlaskit/commits/8aa1bfe))
+* feature; dAC story with renderer (issues closed: ed-2530) ([005bdba](https://bitbucket.org/atlassian/atlaskit/commits/005bdba))
+* feature; render single link card with square appearance ([8197e31](https://bitbucket.org/atlassian/atlaskit/commits/8197e31))
+* feature; add title/user in applicationCard (issues closed: ed-2532) ([a1c5790](https://bitbucket.org/atlassian/atlaskit/commits/a1c5790))
+
+
+* feature; added Confluence transformer (issues closed: ed-1679) ([b9d48b3](https://bitbucket.org/atlassian/atlaskit/commits/b9d48b3))
+* bug fix; pasting plain text in code block should not create a paragraph ([e7b133e](https://bitbucket.org/atlassian/atlaskit/commits/e7b133e))
+* bug fix; should not autoformat __ or _ in mid words. ([49a4df4](https://bitbucket.org/atlassian/atlaskit/commits/49a4df4))
+* bug fix; fix for issue mention mark not removed even after whole text of mention is removed. ([ed7b58d](https://bitbucket.org/atlassian/atlaskit/commits/ed7b58d))
+* bug fix; should not convert "\`some_variables_" to em ([2eaf1f6](https://bitbucket.org/atlassian/atlaskit/commits/2eaf1f6))
+* bug fix; ignore markdown if have a backtick in the front. ([0f921b1](https://bitbucket.org/atlassian/atlaskit/commits/0f921b1))
+## 46.11.1 (2017-08-31)
+
+* bug fix; removed relative links and added tests ([9c07ea7](https://bitbucket.org/atlassian/atlaskit/commits/9c07ea7))
+
+* bug fix; remove auto formatting of relative links (issues closed: ed-2582) ([a36f9dc](https://bitbucket.org/atlassian/atlaskit/commits/a36f9dc))
 ## 46.11.0 (2017-08-30)